Make sure marker is valid before decompressing JPEG tiles
As noted in the comment, missing tiles can be recorded as an offset in a file that does not contain pixel data.
Checking the marker before decompressing is faster as it prevents the JPEGCodec from trying to read the whole
to find a valid marker.  See:

https://github.com/ome/ome-codecs/blob/v0.2.5/src/main/java/ome/codecs/JPEGCodec.java#L108

This also prevents the codec from throwing an exception when it reaches the end of the file or fails to
decompress a valid image.

pyramidgen handled this case by catching Throwable and logging the number of failed tiles.