Re-throw exceptions
Exceptions that are logged at ERROR are not currently
leading to a failure of `Converter.main`. This commit
unwraps thrown exceptions to give feedback to the user.
A breaking alternative would be to introduce return codes
into the currently `void` methods.

This commit was motivated by N5ZarrWriter throwing exceptions
when converting 512x512 (fake) images.