Add tests for int32/uint32 pixel types and fix parameterized tests overall
Pixel type tests are now parameterized, since the same logic is used for
float, double, int32, and uint32.  In changing the pixel type tests,
it became clear that parameterized tests were not being run at all.
This required changes to the JUnit configuration and the downsampling test
to make the build pass as expected.