MRXS: make fill value configurable
Default is now 0 for fluorescence and 255 for brightfield.  For simplicity and speed, the fill value is
applied to all bytes.  We could allow a pixel value instead (so you could specify e.g. magenta for brightfield),
but that's a little more work.

The fill value set with setFillValue(Byte) is intentionally retained after calling close(...).