Use ImageRef to assign each series to a WellSample
The previous codepath was relying on the assumption that the index of
each WellSample could be mapped to each series via its index. While this
applies for the majority of the HCS readers, it is possible to craft an
OME-XML more liberally according to the specification.
This commit proposes to update the test by using WellSamples.ImageRef
and Image.ID as the absolute source of truth.