Update progress listener API to provide tile counts at each start notification
This should make it easier for a listener implementation to use a single progress bar
for the whole conversion, instead of one progress bar per resolution.

Updating the notifications in Converter requires pre-calculating tile and resolution counts
before starting to write tiles, but that also means failing faster if the pixel type and
downsampling (or other options) are incompatible.