MCD: calculate channel count based on metadata, not pixel bytes
In some cases, the number of pixel bytes stored for a particular acquisition
may not match the expected number of bytes, so calculating the channel count
based on the stored number of pixel bytes will cause problems.

Instead of throwing an exception, this sets the channel count to the
number of channels linked to the acquisition, and logs that
some image data may be blank. This more closely matches the behavior
of MCDViewer (the vendor software).